refactor(murmur): split login.rs tests into login/tests.rs
login.rs was 804 lines against this repo's 800-line rule (CLAUDE.md rule 4),
and upcoming tasks add more to this area. Follows the precedent one entry
over in the same directory: fleet_rail.rs + fleet_rail/tests.rs.

Pure move: the #[cfg(test)] mod tests { .. } body relocates verbatim to
login/tests.rs (de-indented, `use super::*;` retained), declared from
login.rs via #[path]. No logic changed; the only diff inside the moved
body is two rustfmt line-wrap adjustments triggered by the reduced
indentation depth.
